% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSimulationParameters.R
\name{generateSimulationParameters}
\alias{generateSimulationParameters}
\title{A function to quickly generate inputs used to run \code{simulateTrip}.}
\usage{
generateSimulationParameters()
}
\value{
A list with the following properties, also lists: \code{baseFuel},
 \code{solarFlares}, \code{asteroids}.
}
\description{
\code{generateSimulationParameters} generates a set of inputs that are to be
used to run \code{simulateTrip}.
}
