% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateTrip.R
\name{simulateTrip}
\alias{simulateTrip}
\title{A function for running an iteration of a monte carlo simulation.}
\usage{
simulateTrip(baseFuel, solarFlares, asteroids)
}
\arguments{
\item{baseFuel}{A list with the keys `copula`, `meanlog`, and `sdlog`. The `copula`
value must be a numeric specifying the parameter of a Clayton copula. The `meanlog`
and `sdlog` keys define a lognormal distribution that models the base amount of
fuel needed for the journey.}

\item{solarFlares}{A list with the keys `fuel`, `lambda`, `scale`, and `shape`.
`fuel` is a list with `meanlog` and `sdlog`. These define a lognormal distribution
that models the amount of additional fuel needed due to a solar flare. `lambda`
characterizes a poisson distribution that models the number of solar flares
occurring on the trip. The `scale` and `shape` parameters define a gamma
distribution that models the intensity of solar flares.}

\item{asteroids}{A list with the keys `fuel` and `lambda`. `fuel` is a list with
the keys `mean` and `sd` which define a normal distribution that models the
additional fuel needed due to avoiding an asteroid. `lambda` characterizes a
poisson distribution that models the number of asteroids occurring on the trip.}
}
\value{
A list with the following properties: `numSolarFlares`, `flareSurface`,
  `baseFuel`, `flareAdditionalFuel`, `asteroidSegments`, `asteroidsPerSegment`,
  `asteroidAdditionalFuel`, and `totalFuel`.
}
\description{
`simulateTrip` takes a set of information regarding asteroids, solar flares, and
fuel requirements that
}
